Fix tabs jumping around
By default tmux displays a * after the name of the current tab, and a -
after the name of the previous tab, but nothing after the names of other
tabs (this is the #F in the window-status-format and
window-status-current-format settings).

This means that when a window becomes or is no longer either current or
previous the length of its name changes. So the window tabs appear to
"jump around" between tabs.

Fix this by rendering a space, rather than nothing, at the end of the
names of windows that are neither current or previous.