Enable mouse support
This turns on scrolling with the mouse, resizing panes with the mouse,
and clicking on a pane or tab with the mouse to go to it.

You can also click and drag to select text with the mouse, which
actually overrides Gnome Terminal's normal select text and right-click
to open context menu and copy. After selecting text in Tmux you have to
do Ctrl-a y to copy it to the system clipboard.