New status bar and window title
Going back to a "tabs"-style bottom status bar, with the current window
highlighted in blue and the last window in red.

Showing ^A in the status bar when Ctrl-a has been pressed, or "Copy"
when in copy mode, with colours matching the current and last tabs.

Window title is just session name.