Tweak the window title format
As the terminal window's title use either the name of the current tmux
window (if it has been manually named) or the name of the command
running in the current tab.

If there is more than one window then also show the current window
number and total number of windows (e.g. "1/2"). This is intended to
enable tmux to be used with no status bar, yet still know where you are
if there are multiple windows.